+++ apps/scan/src/app/_components/health/dot.tsx
