+++ apps/scan/src/app/_components/health/indicator.tsx
