+++ apps/scan/src/app/_components/health/types.ts
