+++ apps/scan/src/app/_components/health/utils.ts
